Migrated editor inputs to non-angular JS
