+++ resources/assets/js/components/index.js
@@ -11,6 +11,8 @@ let componentMapping = {
     'sidebar': require('./sidebar'),
     'page-picker': require('./page-picker'),
     'page-comments': require('./page-comments'),
+    'wysiwyg-editor': require('./wysiwyg-editor'),
+    'markdown-editor': require('./markdown-editor'),
 };
 
 window.components = {};
