+++ resources/assets/js/components/wysiwyg-editor.js
@@ -1,0 +1,11 @@
+class WysiwygEditor {
+
+    constructor(elem) {
+        this.elem = elem;
+        this.options = require("../pages/page-form");
+        tinymce.init(this.options);
+    }
+
+}
+
+module.exports = WysiwygEditor;
